+++ src/analysis/formant/simplelp.cpp
@@ -4,21 +4,21 @@
 using namespace Analysis::Formant;
 using Analysis::FormantResult;
 
-FormantResult SimpleLP::solve(const float *lpc, int lpcOrder, float sampleRate)
+FormantResult SimpleLP::solve(const double *lpc, int lpcOrder, double sampleRate)
 {
-    std::vector<float> polynomial(lpcOrder + 1);
+    std::vector<double> polynomial(lpcOrder + 1);
     polynomial[0] = 1.0f;
     std::copy(lpc, lpc + lpcOrder, std::next(polynomial.begin()));
     
-    std::vector<std::complex<float>> roots = findRoots(polynomial);
+    std::vector<std::complex<double>> roots = findRoots(polynomial);
 
     FormantResult result;
 
     for (const auto& z : roots) {
         if (z.imag() < 0) continue;
 
-        float r = std::abs(z);
-        float phi = std::arg(z);
+        double r = std::abs(z);
+        double phi = std::arg(z);
 
         if (r > 1.0f) {
             continue;
