+++ src/analysis/util/find_peaks.cpp
@@ -3,30 +3,30 @@
 #include <algorithm>
 #include <cmath>
 
-static void diff(std::vector<float> in, std::vector<float>& out)
+static void diff(std::vector<double> in, std::vector<double>& out)
 {
-    out = std::vector<float>(in.size()-1);
+    out = std::vector<double>(in.size()-1);
 
     for(int i=1; i<in.size(); ++i)
         out[i-1] = in[i] - in[i-1];
 }
 
-static void vectorProduct(std::vector<float> a, std::vector<float> b, std::vector<float>& out)
+static void vectorProduct(std::vector<double> a, std::vector<double> b, std::vector<double>& out)
 {
-    out = std::vector<float>(a.size());
+    out = std::vector<double>(a.size());
 
     for(int i=0; i<a.size(); ++i)
         out[i] = a[i] * b[i];
 }
 
-static void findIndicesLessThan(std::vector<float> in, float threshold, std::vector<int>& indices)
+static void findIndicesLessThan(std::vector<double> in, double threshold, std::vector<int>& indices)
 {
     for(int i=0; i<in.size(); ++i)
         if(in[i]<threshold)
             indices.push_back(i+1);
 }
 
-static void selectElements(std::vector<float> in, std::vector<int> indices, std::vector<float>& out)
+static void selectElements(std::vector<double> in, std::vector<int> indices, std::vector<double>& out)
 {
     for(int i=0; i<indices.size(); ++i)
         out.push_back(in[indices[i]]);
@@ -38,7 +38,7 @@ static void selectElements(std::vector<int> in, std::vector<int> indices, std::v
         out.push_back(in[indices[i]]);
 }
 
-static void signVector(std::vector<float> in, std::vector<int>& out, int sign)
+static void signVector(std::vector<double> in, std::vector<int>& out, int sign)
 {
     out = std::vector<int>(in.size());
 
@@ -53,31 +53,31 @@ static void signVector(std::vector<float> in, std::vector<int>& out, int sign)
     }
 }
 
-std::vector<int> Analysis::findPeaks(const float *data, int length, int sign)
+std::vector<int> Analysis::findPeaks(const double *data, int length, int sign)
 {
-    std::vector<float> x0(data, std::next(data, length));
+    std::vector<double> x0(data, std::next(data, length));
     std::vector<int> peakInds;
 
     int minIdx = distance(x0.begin(), min_element(x0.begin(), x0.end()));
     int maxIdx = distance(x0.begin(), max_element(x0.begin(), x0.end()));
 
-    float sel = (x0[maxIdx]-x0[minIdx])/4.0;
+    double sel = (x0[maxIdx]-x0[minIdx])/4.0;
 
     int len0 = x0.size();
 
-    std::vector<float> dx;
+    std::vector<double> dx;
     diff(x0, dx);
     replace(dx.begin(), dx.end(), 0.0f, -2.2204e-16f);
-    std::vector<float> dx0(dx.begin(), dx.end()-1);
-    std::vector<float> dx1(dx.begin()+1, dx.end());
-    std::vector<float> dx2;
+    std::vector<double> dx0(dx.begin(), dx.end()-1);
+    std::vector<double> dx1(dx.begin()+1, dx.end());
+    std::vector<double> dx2;
 
     vectorProduct(dx0, dx1, dx2);
 
     std::vector<int> ind;
     findIndicesLessThan(dx2, 0, ind); // Find where the derivative changes sign
     
-    std::vector<float> x;
+    std::vector<double> x;
 
     std::vector<int> indAux(ind.begin(), ind.end());
     selectElements(x0, indAux, x);
@@ -89,21 +89,21 @@ std::vector<int> Analysis::findPeaks(const float *data, int length, int sign)
     ind.insert(ind.end(), len0);
 
     int minMagIdx = distance(x.begin(), min_element(x.begin(), x.end()));
-    float minMag = x[minMagIdx];
-    float leftMin = minMag;
+    double minMag = x[minMagIdx];
+    double leftMin = minMag;
     int len = x.size();
 
     if(len>2)
     {
-        float tempMag = minMag;
+        double tempMag = minMag;
         bool foundPeak = false;
         int ii;
 
         // Deal with first point a little differently since tacked it on
         // Calculate the sign of the derivative since we tacked the first
         //  point on it does not neccessarily alternate like the rest.
-        std::vector<float> xSub0(x.begin(), x.begin()+3);//tener cuidado substd::vector
-        std::vector<float> xDiff;//tener cuidado substd::vector
+        std::vector<double> xSub0(x.begin(), x.begin()+3);//tener cuidado substd::vector
+        std::vector<double> xDiff;//tener cuidado substd::vector
         diff(xSub0, xDiff);
 
         std::vector<int> signDx;
@@ -133,9 +133,9 @@ std::vector<int> Analysis::findPeaks(const float *data, int length, int sign)
         else
             ii = 1;
 
-        float maxPeaks = ceil((float)len/2.0);
+        double maxPeaks = ceil((double)len/2.0);
         std::vector<int> peakLoc(maxPeaks,0);
-        std::vector<float> peakMag(maxPeaks,0.0);
+        std::vector<double> peakMag(maxPeaks,0.0);
         int cInd = 1;
         int tempLoc;
     
@@ -198,7 +198,7 @@ std::vector<int> Analysis::findPeaks(const float *data, int length, int sign)
         {            
             std::vector<int> peakLocTmp(peakLoc.begin(), peakLoc.begin()+cInd-1);
             selectElements(ind, peakLocTmp, peakInds);
-            //peakMags = std::vector<float>(peakLoc.begin(), peakLoc.begin()+cInd-1);
+            //peakMags = std::vector<double>(peakLoc.begin(), peakLoc.begin()+cInd-1);
         }
     }
